+++ src/main/java/edu/ie3/simona/api/data/model/ev/EvcsModel.java
@@ -1,0 +1,34 @@
+/*
+ * © 2025. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+ */
+
+package edu.ie3.simona.api.data.model.ev;
+
+import java.util.UUID;
+import javax.measure.quantity.Power;
+import tech.units.indriya.ComparableQuantity;
+
+public interface EvcsModel {
+
+  /**
+   * @return the uuid of this evcs
+   */
+  UUID getUuid();
+
+  /**
+   * @return the id of this evcs
+   */
+  String getId();
+
+  /**
+   * @return the maximum AC apparent power rating of this evcs
+   */
+  ComparableQuantity<Power> getSRated();
+
+  /**
+   * @return the rated power factor (cos φ) of this evcs
+   */
+  double getCosPhiRated();
+}